dft: consequences to renaming all the dft files
