+++ doc/sphinxman/CMakeLists.txt
@@ -232,7 +232,7 @@ if(PERL_FOUND AND SPHINX_FOUND AND SPHINX_STUFF_FOUND)
                 ${CCBD}/source/autodoc_dft_dhybrid.rst
                 #${CCBD}/source/autodoc_dft_scsdhybrid.rst  # commented until populated again to appease Sphinx
                 ${CCBD}/source/autodoc_dft_disp.rst
-        DEPENDS ${psi4_SOURCE_DIR}/psi4/driver/procrouting/dft_funcs/superfuncs.py
+        DEPENDS ${psi4_SOURCE_DIR}/psi4/driver/procrouting/dft/superfunctionals.py
                 ${psi4_SOURCE_DIR}/tests/docs-dft/input.dat
                 ${STAGED_INSTALL_PREFIX}/${CMAKE_INSTALL_BINDIR}/psi4
         COMMENT "Autodoc functionals")
