+++ psi4/driver/procrouting/__init__.py
@@ -29,7 +29,7 @@
 from .proc_table import procedures, hooks, energy_only_methods
 from .proc import scf_helper, scf_wavefunction_factory
 from .empirical_dispersion import EmpiricalDispersion
-from . import dft_funcs
+from . import dft
 from . import response
 from . import scf_proc
 from . import libcubeprop
