+++ psi4/driver/procrouting/dft/__init__.py
@@ -26,5 +26,5 @@
 # @END LICENSE
 #
 
-from .superfuncs import *
-from .dict_builder import *
+from .superfunctionals import *
+from .dft_builder import *
