+++ psi4/driver/procrouting/proc_util.py
@@ -34,8 +34,7 @@ import numpy as np
 from psi4 import core
 from psi4.driver import p4util
 from psi4.driver.p4util.exceptions import *
-from psi4.driver.procrouting.dft_funcs import functionals
-from psi4.driver.procrouting.dft_funcs import build_superfunctional_from_dictionary
+from psi4.driver.procrouting.dft import functionals, build_superfunctional_from_dictionary
 
 def scf_set_reference_local(name, is_dft=False):
     """
