+++ tests/docs-dft/input.dat
@@ -1,8 +1,8 @@
 #! This test is used to construct the documentation;
 #! it is not suitable for emulation by users. !nosample
 
-from psi4.driver.procrouting.dft_funcs import functionals
-from psi4.driver.procrouting.dft_funcs import build_superfunctional_from_dictionary
+from psi4.driver.procrouting.dft import functionals
+from psi4.driver.procrouting.dft import build_superfunctional_from_dictionary
 
 funcdict = {}
